re-organize simulator
